[DiskBBQ] Merge SampleReader into FloatVectorValuesSlice
